+++ internal/container/rest.go
@@ -1,17 +1,12 @@
 package container
 
-import "github.com/gin-gonic/gin"
+import (
+	"github.com/DucTran999/auth-service/internal/handler/rest"
+	"github.com/gin-gonic/gin"
+)
 
 type HealthHandler interface {
 	CheckLiveness(ctx *gin.Context)
-}
-
-type AuthHandler interface {
-	LoginAccount(ctx *gin.Context)
-	LoginAccountJWT(ctx *gin.Context)
-	LogoutAccount(ctx *gin.Context)
-	RefreshToken(ctx *gin.Context)
-	LogoutAccountJWT(ctx *gin.Context)
 }
 
 type AccountHandler interface {
@@ -20,15 +15,17 @@ type AccountHandler interface {
 }
 
 type RestHandler struct {
-	AuthHandler
+	rest.SessionAuthHandler
+	rest.JWTAuthHandler
 	AccountHandler
 	HealthHandler
 }
 
 func (c *Container) initRestHandler() {
 	c.RestHandler = &RestHandler{
-		AuthHandler:    c.handlers.auth,
-		AccountHandler: c.handlers.account,
-		HealthHandler:  c.handlers.health,
+		JWTAuthHandler:     c.handlers.jwtAuth,
+		SessionAuthHandler: c.handlers.sessionAuth,
+		AccountHandler:     c.handlers.account,
+		HealthHandler:      c.handlers.health,
 	}
 }
