+++ internal/container/usecase.go
@@ -3,12 +3,14 @@ package container
 import (
 	"github.com/DucTran999/auth-service/internal/handler/background"
 	"github.com/DucTran999/auth-service/internal/usecase"
+	"github.com/DucTran999/auth-service/internal/usecase/auth"
 	"github.com/DucTran999/auth-service/internal/usecase/port"
 	"github.com/DucTran999/auth-service/internal/usecase/shared"
 )
 
 type useCases struct {
-	auth              port.AuthUseCase
+	jwtAuth           port.JWTAuthUsecase
+	sessionAuth       port.AuthUseCase
 	account           port.AccountUseCase
 	restSession       port.SessionUsecase
 	backgroundSession background.SessionUsecase
@@ -24,20 +26,21 @@ func (c *Container) initUseCases() {
 		c.repositories.account,
 	)
 
-	authUC := usecase.NewAuthUseCase(
+	sessionAuthUC := auth.NewAuthUseCase(
 		c.Hasher,
-		c.Signer,
 		c.Cache,
 		accountVerifier,
 		c.repositories.account,
 		c.repositories.session,
 	)
+	jwtAuthUC := auth.NewAuthJWTUsecase(c.Signer, c.Cache, accountVerifier)
 
 	sessionUC := usecase.NewSessionUC(c.Cache, c.repositories.session)
 
 	c.useCases = &useCases{
 		account:           accountUC,
-		auth:              authUC,
+		jwtAuth:           jwtAuthUC,
+		sessionAuth:       sessionAuthUC,
 		restSession:       sessionUC,
 		backgroundSession: sessionUC,
 	}
