+++ internal/usecase/port/session_auth_usecase.go
@@ -17,10 +17,4 @@ type AuthUseCase interface {
 	// It removes the session from cache (best-effort) and marks it as expired in the database.
 	// Returns an error only if the database update fails.
 	Logout(ctx context.Context, sessionID string) error
-
-	LoginJWT(ctx context.Context, input dto.LoginJWTInput) (*dto.TokenPairs, error)
-
-	RefreshToken(ctx context.Context, refreshToken string) (*dto.TokenPairs, error)
-
-	RevokeRefreshToken(ctx context.Context, refreshToken string) error
 }
